+++ source/B2B.Transactions.CimMessageAdapter.Tests/MessageReceiverTests.cs
@@ -16,6 +16,7 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using B2B.CimMessageAdapter;
+using B2B.CimMessageAdapter.Schema;
 using MarketRoles.B2B.CimMessageAdapter.IntegrationTests.Stubs;
 using Xunit;
