+++ source/B2B.Transactions.CimMessageAdapter/MessageReceiver.cs
@@ -18,6 +18,7 @@ using System.IO;
 using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.Schema;
+using B2B.CimMessageAdapter.Schema;
 
 namespace B2B.CimMessageAdapter
 {
